chore: unified the naming for demand and capacity notification
