+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/domain/model/OwnDemandAndCapacityNotification.java
@@ -18,15 +18,17 @@ under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package org.eclipse.tractusx.puris.backend.notification.domain.repository;
+package org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model;
 
-import java.util.UUID;
+import jakarta.persistence.Entity;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
 
-import org.eclipse.tractusx.puris.backend.notification.domain.model.ReportedDemandAndCapacityNotification;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReportedDemandAndCapacityNotificationRepository extends JpaRepository<ReportedDemandAndCapacityNotification, UUID>{
+@Entity
+@SuperBuilder
+@NoArgsConstructor
+@ToString(callSuper = true)
+public class OwnDemandAndCapacityNotification extends DemandAndCapacityNotification {
 
 }
