+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/domain/repository/ReportedDemandAndCapacityNotificationRepository.java
@@ -18,17 +18,15 @@ under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package org.eclipse.tractusx.puris.backend.notification.domain.model;
+package org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.repository;
 
-import jakarta.persistence.Entity;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
+import java.util.UUID;
 
-@Entity
-@SuperBuilder
-@NoArgsConstructor
-@ToString(callSuper = true)
-public class OwnDemandAndCapacityNotification extends DemandAndCapacityNotification {
+import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model.ReportedDemandAndCapacityNotification;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReportedDemandAndCapacityNotificationRepository extends JpaRepository<ReportedDemandAndCapacityNotification, UUID>{
 
 }
