+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/logic/service/OwnDemandAndCapacityNotificationService.java
@@ -18,31 +18,29 @@ under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package org.eclipse.tractusx.puris.backend.notification.logic.service;
+package org.eclipse.tractusx.puris.backend.demandandcapacitynotification.logic.service;
 
 import java.util.List;
 
+import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.model.OwnDemandAndCapacityNotification;
+import org.eclipse.tractusx.puris.backend.demandandcapacitynotification.domain.repository.OwnDemandAndCapacityNotificationRepository;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialPartnerRelationService;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.service.PartnerService;
-import org.eclipse.tractusx.puris.backend.notification.domain.model.ReportedDemandAndCapacityNotification;
-import org.eclipse.tractusx.puris.backend.notification.domain.repository.ReportedDemandAndCapacityNotificationRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ReportedDemandAndCapacityNotificationService extends DemandAndCapacityNotificationService<ReportedDemandAndCapacityNotification, ReportedDemandAndCapacityNotificationRepository> {
+public class OwnDemandAndCapacityNotificationService extends DemandAndCapacityNotificationService<OwnDemandAndCapacityNotification, OwnDemandAndCapacityNotificationRepository>{
 
-    public ReportedDemandAndCapacityNotificationService(ReportedDemandAndCapacityNotificationRepository reportedNotificationRepository,
-            PartnerService partnerService, MaterialPartnerRelationService mpr) {
-        super(reportedNotificationRepository, partnerService, mpr);
+    public OwnDemandAndCapacityNotificationService(OwnDemandAndCapacityNotificationRepository ownNotificationRepository, PartnerService partnerService, MaterialPartnerRelationService mpr) {
+        super(ownNotificationRepository, partnerService, mpr);
     }
 
-    public List<ReportedDemandAndCapacityNotification> findAllByPartnerBpnl(String bpnl) {
-        return repository.findAll().stream().filter(notification -> notification.getPartner().getBpnl().equals(bpnl))
-                .toList();
+    public List<OwnDemandAndCapacityNotification>  findAllByPartnerBpnl(String bpnl) {
+        return repository.findAll().stream().filter(notification -> notification.getPartner().getBpnl().equals(bpnl)).toList();
     }
 
     @Override
-    public boolean validate(ReportedDemandAndCapacityNotification notification) {
+    public boolean validate(OwnDemandAndCapacityNotification notification) {
         return notification.getPartner() != null &&
                 notification.getText() != null &&
                 notification.getLeadingRootCause() != null &&
@@ -54,7 +52,7 @@ public class ReportedDemandAndCapacityNotificationService extends DemandAndCapac
                 validateSites(notification);
     }
 
-    public boolean validateMaterials(ReportedDemandAndCapacityNotification notification) {
+    public boolean validateMaterials(OwnDemandAndCapacityNotification notification) {
         if (notification.getMaterials() == null || notification.getMaterials().isEmpty()) {
             return true;
         }
@@ -63,15 +61,15 @@ public class ReportedDemandAndCapacityNotificationService extends DemandAndCapac
             notification.getMaterials().contains(mpr.getMaterial())).count() == notification.getMaterials().size();
     }
 
-    public boolean validateSites(ReportedDemandAndCapacityNotification notification) {
+    public boolean validateSites(OwnDemandAndCapacityNotification notification) {
         return (
-            notification.getAffectedSitesSender() == null ||
-            notification.getAffectedSitesSender().isEmpty() ||
-            notification.getAffectedSitesSender().stream().allMatch(site -> notification.getPartner().getSites().contains(site))
-        ) && (
             notification.getAffectedSitesRecipient() == null ||
             notification.getAffectedSitesRecipient().isEmpty() ||
-            notification.getAffectedSitesRecipient().stream().allMatch(site -> partnerService.getOwnPartnerEntity().getSites().contains(site))
+            notification.getAffectedSitesRecipient().stream().allMatch(site -> notification.getPartner().getSites().contains(site))
+        ) && (
+            notification.getAffectedSitesSender() == null ||
+            notification.getAffectedSitesSender().isEmpty() ||
+            notification.getAffectedSitesSender().stream().allMatch(site -> partnerService.getOwnPartnerEntity().getSites().contains(site))
         );
     }
 }
